Add Location

+++ app/models/location.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, ForeignKey, String, Float
+from sqlalchemy import Column, ForeignKey, Numeric, String, Float
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import relationship
 
@@ -9,7 +9,7 @@ class Location(BaseModel):
     __tablename__ = "locations"
 
     name = Column(String, nullable=False)
-    price = Column(Float)
+    price = Column(Numeric(10, 2))  # Changed Float to Numeric for precision
     parent_id = Column(UUID(as_uuid=True), ForeignKey("locations.id"))
 
     # Define the children relationship (one-to-many)
